Use undo-based rollback to clean up files on abort.
Use the undo log instead of "pendingDeletes" to track new files that must
be deleted if the transaction rolls back.  This way, if you kill -9 the
backend after CREATE TABLE but before COMMIT, the orphaned files will be
cleaned up after crash recovery.

This requires logging and flushing a new XLOG_SMGR_PRECREATE record
before creating relfiles, so that we know the undo records will be
recreated if we crash and recover.  During undo action execution,
we log a new XLOG_SMGR_DROP record that will unlink the file on the
standby.

Refactor undoaction.c so that it dispatches to undo handlers via
RmgrTable.  Define smgr_undo() for the orphaned file cleanup routine.

Add urec_rmid to all undo record headers so that we know which
RMGR will execute the record.

TODO:  Don't put RMGR IDs onto records, when we can derive them from
table access managers?

TODO:  This can leak empty files if you crash after a checkpoint!

TODO:  This patch undoes the AtAbort stuff from an earlier patch;
sort that out, it doesn't belong here.

TODO:  There is one regression test that gets an assertion failure:
rollback to savepoint where a cursor is in mid-scan, causes "buffer
is pinned in InvalidateBuffer", because we are trying to run
DropRelFileNodesAllBuffers() on a pinned buffer, when unlinking
a newly created relation.

Author: Thomas Munro